+++ codegen/example/src-gen/main/java/org/fuin/ddd4j/codegen/example/TheRootCreateCommand.java
@@ -5,6 +5,7 @@ import jakarta.json.bind.annotation.JsonbProperty;
 import jakarta.json.bind.annotation.JsonbTypeAdapter;
 import javax.annotation.Nullable;
 import org.fuin.ddd4j.core.EntityIdPath;
+import org.fuin.ddd4j.core.EventId;
 import org.fuin.ddd4j.core.EventType;
 import org.fuin.cqrs4j.jsonb.AbstractAggregateCommand;
 import org.fuin.ddd4j.jsonb.AbstractDomainEvent;
@@ -13,6 +14,7 @@ import org.fuin.esc.api.SerializedDataType;
 import org.fuin.objects4j.common.Contract;
 
 import java.io.Serial;
+import java.time.ZonedDateTime;
 
 import javax.annotation.concurrent.Immutable;
 
@@ -178,6 +180,12 @@ public final class TheRootCreateCommand extends AbstractAggregateCommand<TheRoot
          */
         public TheRootCreateCommand build() {
             ensureBuildableAbstractAggregateCommand();
+            if (delegate.getEventId() == null) {
+                this.eventId(new EventId());
+            }
+            if (delegate.getEventTimestamp() == null) {
+                this.timestamp(ZonedDateTime.now());
+            }
             final TheRootCreateCommand result = delegate;
             delegate = new TheRootCreateCommand();
             resetAbstractAggregateCommand(delegate);