+++ codegen/example/src-gen/main/java/org/fuin/ddd4j/codegen/example/TheRootCreatedEvent.java
@@ -5,6 +5,7 @@ import jakarta.json.bind.annotation.JsonbProperty;
 import jakarta.json.bind.annotation.JsonbTypeAdapter;
 import javax.annotation.Nullable;
 import org.fuin.ddd4j.core.EntityIdPath;
+import org.fuin.ddd4j.core.EventId;
 import org.fuin.ddd4j.core.EventType;
 import org.fuin.ddd4j.jsonb.AbstractDomainEvent;
 import org.fuin.esc.api.HasSerializedDataTypeConstant;
@@ -12,6 +13,7 @@ import org.fuin.esc.api.SerializedDataType;
 import org.fuin.objects4j.common.Contract;
 
 import java.io.Serial;
+import java.time.ZonedDateTime;
 
 import javax.annotation.concurrent.Immutable;
 
@@ -177,6 +179,12 @@ public final class TheRootCreatedEvent extends AbstractDomainEvent<TheRootId> {
          */
         public TheRootCreatedEvent build() {
             ensureBuildableAbstractDomainEvent();
+            if (delegate.getEventId() == null) {
+                this.eventId(new EventId());
+            }
+            if (delegate.getEventTimestamp() == null) {
+                this.timestamp(ZonedDateTime.now());
+            }
             final TheRootCreatedEvent result = delegate;
             delegate = new TheRootCreatedEvent();
             resetAbstractDomainEvent(delegate);