+++ codegen/processor/src/main/resources/EventVO.java
@@ -20,6 +20,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 #end
 import javax.annotation.Nullable;
 import org.fuin.ddd4j.core.EntityIdPath;
+import org.fuin.ddd4j.core.EventId;
 import org.fuin.ddd4j.core.EventType;
 #if($jsonb)
 import org.fuin.ddd4j.jsonb.AbstractDomainEvent;
@@ -38,6 +39,7 @@ import ${idClass};
 #end
 
 import java.io.Serial;
+import java.time.ZonedDateTime;
 
 import javax.annotation.concurrent.Immutable;
 
@@ -159,6 +161,12 @@ public final class ${class} extends AbstractDomainEvent<${idClass.simpleName}> {
          */
         public ${class} build() {
             ensureBuildableAbstractDomainEvent();
+            if (delegate.getEventId() == null) {
+                this.eventId(new EventId());
+            }
+            if (delegate.getEventTimestamp() == null) {
+                this.timestamp(ZonedDateTime.now());
+            }
             final ${class} result = delegate;
             delegate = new ${class}();
             resetAbstractDomainEvent(delegate);