+++ codegen/processor/src/test/resources/expected/MyCommand/Jaxb.java
@@ -8,6 +8,7 @@ import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import javax.annotation.Nullable;
 import org.fuin.ddd4j.core.EntityIdPath;
+import org.fuin.ddd4j.core.EventId;
 import org.fuin.ddd4j.core.EventType;
 import org.fuin.cqrs4j.jaxb.AbstractAggregateCommand;
 import org.fuin.ddd4j.jaxb.AbstractDomainEvent;
@@ -16,6 +17,7 @@ import org.fuin.esc.api.SerializedDataType;
 import org.fuin.objects4j.common.Contract;
 
 import java.io.Serial;
+import java.time.ZonedDateTime;
 
 import javax.annotation.concurrent.Immutable;
 
@@ -118,6 +120,12 @@ public final class MyCommand extends AbstractAggregateCommand<MyRootId, MyId> {
          */
         public MyCommand build() {
             ensureBuildableAbstractAggregateCommand();
+            if (delegate.getEventId() == null) {
+                this.eventId(new EventId());
+            }
+            if (delegate.getEventTimestamp() == null) {
+                this.timestamp(ZonedDateTime.now());
+            }
             final MyCommand result = delegate;
             delegate = new MyCommand();
             resetAbstractAggregateCommand(delegate);