+++ codegen/processor/src/test/resources/expected/MyEvent/Jaxb.java
@@ -8,6 +8,7 @@ import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import javax.annotation.Nullable;
 import org.fuin.ddd4j.core.EntityIdPath;
+import org.fuin.ddd4j.core.EventId;
 import org.fuin.ddd4j.core.EventType;
 import org.fuin.ddd4j.jaxb.AbstractDomainEvent;
 import org.fuin.esc.api.HasSerializedDataTypeConstant;
@@ -15,6 +16,7 @@ import org.fuin.esc.api.SerializedDataType;
 import org.fuin.objects4j.common.Contract;
 
 import java.io.Serial;
+import java.time.ZonedDateTime;
 
 import javax.annotation.concurrent.Immutable;
 
@@ -119,6 +121,12 @@ public final class MyEvent extends AbstractDomainEvent<MyId> {
          */
         public MyEvent build() {
             ensureBuildableAbstractDomainEvent();
+            if (delegate.getEventId() == null) {
+                this.eventId(new EventId());
+            }
+            if (delegate.getEventTimestamp() == null) {
+                this.timestamp(ZonedDateTime.now());
+            }
             final MyEvent result = delegate;
             delegate = new MyEvent();
             resetAbstractDomainEvent(delegate);