+++ release-notes.md
@@ -7,6 +7,7 @@
 - New [annotation](codegen/api/src/main/java/org/fuin/ddd4j/codegen/api/CommandVO.java) and code generation template for commands
 - Added missing annotation "@HasPublicStaticIsValidMethod" to generated entity and aggregate identifiers
 - Fixed bug: Code was not generated in the package configured in the annotation.
+- Added default value in generated command and event builders for eventId and timestamp
 
 ## 0.7.0
 - Added [Jackson](jackson) module